requirements.txt
amqp==2.2.2
bleach==1.4.3
certifi==2018.4.16
chardet==3.0.4
diff-match-patch==20121119
dj-database-url==0.5.0
Django==2.0.2
django-filter==1.1.0
django-graphiql==0.4.4
django-heroku==0.3.1
django-import-export==1.0.1
django-seed==0.1.8
django-tenant-schemas==1.9.0
django-widget-tweaks==1.4.1
django-widgets-improved==1.5.0
djangorestframework==3.7.7
et-xmlfile==1.0.1
Faker==0.8.17
gunicorn==19.9.0
html5lib==0.9999999
idna==2.6
jdcal==1.4
odfpy==1.3.6
openpyxl==2.5.4
Pillow==5.2.0
pipenv==2018.7.1
psycopg2==2.7.5
python-dateutil==2.7.3
python-decouple==3.1
pytz==2018.5
PyYAML==3.13
requests==2.18.4
six==1.11.0
tablib==0.12.1
text-unidecode==1.2
unicodecsv==0.14.1
Unipath==1.0
urllib3==1.22
vine==1.1.4
virtualenv==16.0.0
virtualenv-clone==0.3.0
whitenoise==3.3.1
xlrd==1.1.0
xlwt==1.3.0
